filter, auth, cors, security and jwt

+++ src/main/java/com/mynager/rest/api/services/UserServiceImpl.java
@@ -6,6 +6,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.mynager.rest.api.config.UserSpringSecurity;
+import com.mynager.rest.api.controller.exception.AuthorizationException;
+import com.mynager.rest.api.controller.exception.EmailAlreadyExistsException;
 import com.mynager.rest.api.controller.exception.NotFoundException;
 import com.mynager.rest.api.model.User;
 import com.mynager.rest.api.repository.UserRepository;
@@ -23,16 +26,21 @@ public class UserServiceImpl implements UserService {
 	public User findByEmail(String email) {
 		User user = usRepository.findByEmail(email);
 		if (user == null) {
-			throw new NotFoundException("Ops! User not found.");
+			throw new NotFoundException("User not found.");
 		}
 		return user;
 	}
 
 	@Override
 	public User findById(long id) {
+		UserSpringSecurity userCont = UserSpringSecurityServiceImpl.authenticated();
+		if (userCont == null || id != userCont.getId()) {
+			throw new AuthorizationException("Access Denied.");
+		}
+
 		User user = usRepository.findById(id);
 		if (user == null) {
-			throw new NotFoundException("Ops! User not found.");
+			throw new NotFoundException("User not found.");
 		}
 		return user;
 	}
@@ -40,23 +48,35 @@ public class UserServiceImpl implements UserService {
 	@Override
 	public void save(User user) {
 		user.setId(null);
+		User userEquals = usRepository.findByEmail(user.getEmail());
+		if (userEquals != null) {
+			throw new EmailAlreadyExistsException("Email already exists!");
+		}
 		user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
 		usRepository.save(user);
 	}
-	
+
 	@Override
 	public void update(User user) {
+		UserSpringSecurity userCont = UserSpringSecurityServiceImpl.authenticated();
+		if (userCont == null || user.getId() != userCont.getId()) {
+			throw new AuthorizationException("Access Denied.");
+		}
 		user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
 		usRepository.save(user);
 	}
 
 	@Override
 	public void delete(long id) {
+		UserSpringSecurity userCont = UserSpringSecurityServiceImpl.authenticated();
+		if (userCont == null || id != userCont.getId()) {
+			throw new AuthorizationException("Access Denied.");
+		}
 		usRepository.deleteById(id);
 	}
 
 	@Override
 	public List<User> findAll() {
 		return usRepository.findAll();
-	}	
+	}
 }
